[enh] Rename the package of the EditingContext collaborative support
The package was still named with "project" because it used to be the location
of the ProjectEventProcessor. This concept has been removed a long time ago
to be replaced by the EditingContextEventProcessor.
