+++ CHANGELOG.adoc
@@ -17,7 +17,8 @@
 - https://github.com/eclipse-sirius/sirius-components/issues/878[#878] [core] As specified in the ADR-36, the kind of an object (representation, semantic elements, selection entry of the workbench) is now an URI. All previous kind values have been modified
 - https://github.com/eclipse-sirius/sirius-components/issues/878[#878] [graphql] The GraphQL argument `classId` which appeared on some fields has been replaced by `kind` since it was always the `kind` of an object. Technically, it is always the kind of a semantic element but that may not be the case forever
 - [core] `IEditService.findClass()` has been removed
-- [core] `IRepresentationDescriptionSearchService` has a new `findAll` method to return all the representation descriptions available in a given editing context.
+- [core] `IRepresentationDescriptionSearchService` has a new `findAll` method to return all the representation descriptions available in a given editing context
+- [core] The package containing the concepts related to the editing context has been renamed to remove references to the project
 
 === Dependency update
