+++ backend/sirius-web-spring-collaborative/src/main/java/org/eclipse/sirius/web/spring/collaborative/editingcontext/EditingContextEventProcessorRegistry.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     Obeo - initial API and implementation
  *******************************************************************************/
-package org.eclipse.sirius.web.spring.collaborative.projects;
+package org.eclipse.sirius.web.spring.collaborative.editingcontext;
 
 import java.time.Duration;
 import java.util.List;
